Let whole source be a button
